+++ src/test/java/webserver/controller/UserListControllerTest.java
@@ -10,7 +10,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.web.reactive.function.BodyInserters;
 import webserver.WebServer;
-import webserver.support.CookieParser;
+import http.cookie.CookieParser;
 
 import static org.assertj.core.api.Assertions.assertThat;
