+++ src/test/java/webserver/support/CookieParserTest.java
@@ -1,6 +1,7 @@
 package webserver.support;
 
 import http.cookie.Cookie;
+import http.cookie.CookieParser;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
